requirements.txt
python-dotenv
streamlit
pypdf
langchain
langchain-community
langchain-text-splitters
langchain_openai
langchain_chroma
duckdb
pandas